# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API for Golang Project Ad Mediation.",
        "title": "Ad Mediation Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "luka.kosenina@outlook.com"
        },
        "license": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/adNetworkList": {
            "get": {
                "description": "get ad network list for queried time, current (for today) ad network list is returned if parameter is not provided",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adNetworkList"
                ],
                "summary": "Get ad network list",
                "parameters": [
                    {
                        "type": "string",
                        "format": "date",
                        "description": "provide time to fetch list of ad newtworks for specified time",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AdNetworkList"
                        }
                    }
                }
            },
            "put": {
                "description": "Update or insert ad network list for today",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adNetworkList"
                ],
                "summary": "Upsert ad network list",
                "parameters": [
                    {
                        "description": "Upsert ad network list",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.AdNetworkList"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        }
    },
    "definitions": {
        "models.AdNetwork": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "rank": {
                    "type": "integer"
                }
            }
        },
        "models.AdNetworkList": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AdNetwork"
                    }
                }
            }
        }
    }
}